requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
django-filter==2.2.0
djangorestframework==3.10.1
Markdown==3.1.1
Pillow==5.4.1
pytz==2018.9
